% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/kmeans.R
\docType{methods}
\name{POWER}
\alias{POWER}
\alias{POWER,ANY-method}
\title{Matrix power}
\usage{
POWER(x, ...)

\S4method{POWER}{ANY}(x, k = 1)
}
\arguments{
\item{x}{the matrix}

\item{k}{power value (default k=1)}
}
\value{
a list with the following attributes: input matrix, power value,
 power matrix
}
\description{
This is an implementation of a package where return the power
  between a matrix and a value choose by user.
}
\details{
The implementation admits a matrix and a value.
  In the end, it returns the power between the matrix and the value.
  Matrix, HDF5Matrix and DelayMatrix are not allow and the system returns an error.
}
\references{
https://github.com/mlampros/ClusterR
}
