% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powerRcpp}
\alias{powerRcpp}
\title{Mini_batch}
\usage{
powerRcpp(data, k)
}
\arguments{
\item{data}{numeric matrix or integer matrix or HDF5Matrix}

\item{k}{the number of clusters}

\item{batch_size}{the size of the mini batches}
}
\value{
a list with the following attributes: centroids, WCSS_per_cluster, best_initialization, iters_per_initialization
}
\description{
Mini-batch-k-means for both matrix and HDF5Matrix
}
\details{
This function performs k-means clustering using mini batches.

\strong{kmeans++}: kmeans++ initialization. Reference : http://theory.stanford.edu/~sergei/papers/kMeansPP-soda.pdf AND http://stackoverflow.com/questions/5466323/how-exactly-does-k-means-work

\strong{random}: random selection of data rows as initial centroids
}
\references{
https://github.com/mlampros/ClusterR
}
